% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_genome_size.R
\name{build_lmm_model}
\alias{build_lmm_model}
\title{Build a frequentist random effect model using nested effects}
\usage{
build_lmm_model(size_db, effects)
}
\arguments{
\item{size_db}{Genome size reference database}

\item{effects}{Vector of nested effects to use in the formula e.g. c("family", "genus")}
}
\description{
Build a frequentist random effect model using nested effects
}
