% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_genome_size.R
\name{get_bayes_model}
\alias{get_bayes_model}
\title{Read bayesian model from rds file}
\usage{
get_bayes_model(superkingdom)
}
\arguments{
\item{superkingdom}{Target superkingdom, taxid or name}
}
\description{
Read bayesian model from rds file
}
