% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_genome_size.R
\name{get_genome_size_db_for_lmm}
\alias{get_genome_size_db_for_lmm}
\title{Read genome size database for lmm model}
\usage{
get_genome_size_db_for_lmm(genome_size_db_path = NA)
}
\arguments{
\item{genome_size_db_path}{Path to genome size database file or NA}
}
\description{
Read genome size database for lmm model
}
