% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_genome_size.R
\name{get_refdata}
\alias{get_refdata}
\title{Untar external data}
\usage{
get_refdata(refdata_path)
}
\arguments{
\item{data_path}{Path to tar.gz archive or NA}
}
\description{
Untar external data
}
