% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_genome_size.R
\name{get_taxonomy}
\alias{get_taxonomy}
\title{Read taxonomy database}
\usage{
get_taxonomy(taxonomy_path)
}
\arguments{
\item{taxonomy_path}{Path to taxonomy database file or NA}
}
\description{
Read taxonomy database
}
