% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_genome_size_boxplot}
\alias{plot_genome_size_boxplot}
\title{Plot genome size boxplot}
\usage{
plot_genome_size_boxplot(
  output_table,
  sample_data = NA,
  only_sample = NA,
  PA = F
)
}
\arguments{
\item{output_table}{Result table from estimate_genome_size}

\item{sample_data}{The file or dataframe containing sample information (sample name and count)}

\item{only_sample}{The sample to plot the genome size boxplot for (default: NA, all samples)}

\item{PA}{Use presence/absence data instead of abundance data (all counts are set to 1)}
}
\description{
This function loads a result table from estimate_genome_size and plots a boxplot of estimated genome sizes.
}
