% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_genome_size_per_sample}
\alias{plot_genome_size_per_sample}
\title{Plot genome size per sample}
\usage{
plot_genome_size_per_sample(per_sample_table, order_by_size = TRUE)
}
\arguments{
\item{per_sample_table}{Result table from estimate_genome_size_per_sample()}

\item{order_by_size}{Order samples by estimated genome size (default: TRUE) rather than by sample name}
}
\description{
This function loads a result table from estimate_genome_size_per_sample and plots the estimated mean genome
size of each sample with its interval.
}
