% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_genome_size_tree}
\alias{plot_genome_size_tree}
\title{Plot genome size tree}
\usage{
plot_genome_size_tree(output_table, refdata_path)
}
\arguments{
\item{output_table}{Result table from estimate_genome_size}

\item{refdata_path}{Path to the downloadable archive containing the reference databases and the bayesian models}
}
\description{
This function loads a result table from estimate_genome_size and plots a taxonomic tree with colour-coded estimated genome sizes.
}
