% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_genome_size.R
\name{estimate_genome_size}
\alias{estimate_genome_size}
\title{Infer genome sizes}
\usage{
estimate_genome_size(
  queries,
  format = "csv",
  sep = ",",
  match_column = NA,
  match_sep = ",",
  size_db = NA,
  taxonomy = NA,
  output_format = "input"
)
}
\arguments{
\item{queries}{Queries: path to file or table object}

\item{format}{Query format if in a file ('csv' (default) or 'dada2' (taxonomy table file))}

\item{sep}{If 'csv' format, column separator}

\item{match_column}{If 'csv' format, the column containing match information (with one or several matches)}

\item{match_sep}{If 'csv' format and several matches in match column, separator between matches}

\item{size_db}{Path to the genome size reference database to use if not the default one}

\item{taxonomy}{Path to taxonomy database (NCBI taxdump) to use if not the default one}

\item{output_format}{Format in which the output should be.
Default: "input" a data frame with the same columns as the input,
with the added columns: "estimated_genome_size" and "estimated_genome_size_confidence_interval".
Other formats available: "data.frame", a data frame with the same number of rows as the input, and 2 columns:
"estimated_genome_size" and "estimated_genome_size_confidence_interval".}
}
\description{
This function loads a query file and predicts genome sizes.
}
