% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_genome_size.R
\name{get_refdata}
\alias{get_refdata}
\title{Untar reference data archive containing reference databases and bayesion models}
\usage{
get_refdata(refdata_path)
}
\arguments{
\item{refdata_path}{Path to tar.gz archive}
}
\description{
Untar reference data archive containing reference databases and bayesion models
}
